refactor: Move tag model, repo, view comp to Tags namespace #266
