+++ src/Fan.Blog/Data/BlogEntityModelBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using Fan.Blog.Models;
+using Fan.Blog.Tags;
 using Fan.Data;
 using Microsoft.EntityFrameworkCore;
