+++ src/Fan.Blog/Data/SqlPostRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Fan.Blog.Enums;
 using Fan.Blog.Models;
+using Fan.Blog.Tags;
 using Fan.Data;
 using Microsoft.EntityFrameworkCore;
 using System;
