+++ src/Fan.Blog/ViewModels/BlogPostListViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using Fan.Blog.Enums;
 using Fan.Blog.Models;
 using Fan.Blog.Services;
+using Fan.Blog.Tags;
 using Microsoft.AspNetCore.Http;
 using System.Collections.Generic;
