+++ src/Fan.Blog/ViewModels/BlogPostViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using Fan.Blog.Enums;
 using Fan.Blog.Helpers;
 using Fan.Blog.Models;
+using Fan.Blog.Tags;
 using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
