+++ src/Fan.Web/Themes/Clarity/Views/_ViewImports.cshtml
@@ -1,5 +1,6 @@
 @using Fan.Blog.Enums
 @using Fan.Blog.Models
+@using Fan.Blog.Tags
 @using Fan.Blog.ViewModels
 @using Fan.Blog.Services
 @using Fan.Membership
