+++ test/Fan.Blog.IntegrationTests/SqlTagRepositoryTest.cs
@@ -1,6 +1,7 @@
-﻿using Fan.Blog.IntegrationTests.Base;
-using Fan.Blog.Data;
+﻿using Fan.Blog.Data;
+using Fan.Blog.IntegrationTests.Base;
 using Fan.Blog.Models;
+using Fan.Blog.Tags;
 using System.Linq;
 using Xunit;
